Updated Express
Updated to Express v5 for the 2nd edition of Bootstrapping Microservices.

v5 is currently in beta, by the time v5 is published hopefully it will be out of beta and this code can be updated.

v5 is useful because it supports asynchronous route handlers will simplify code later in the book.